refactor(python-sdk): move template paginator into the template SDK
Align Python with the TypeScript layout: template code no longer lives in the
sandbox modules.

- TemplateInfo -> e2b/template/template_api.py (mirrors templateApi.ts)
- TemplatePaginator -> e2b/template_sync/paginator.py
- AsyncTemplatePaginator -> e2b/template_async/paginator.py
- both extend the shared PaginatorBase directly (drop the redundant
  TemplatePaginatorBase), return [] when exhausted, and raise TemplateException
- template_sync/async main + e2b/__init__ import from the template modules; no
  sandbox imports remain in the template SDK
