+++ packages/python-sdk/e2b/template_sync/main.py
@@ -24,7 +24,7 @@ from e2b.template_sync.build_api import (
     wait_for_build_finish,
 )
 from e2b.template.utils import normalize_build_arguments, read_dockerignore
-from e2b.sandbox_sync.paginator import TemplatePaginator
+from e2b.template_sync.paginator import TemplatePaginator
 
 
 class Template(TemplateBase):
